+++ hardpy/hardpy_panel/frontend/src/hardpy_test_view/SuiteList.tsx
@@ -86,7 +86,12 @@ export interface TestRunI {
 interface Props extends WithTranslation {
   db_state: TestRunI;
   defaultClose: boolean;
-  measurementDisplay?: boolean; 
+  onTestsSelectionChange?: (selectedTests: string[]) => void;
+  selectedTests?: string[];
+  selectionSupported?: boolean;
+  measurementDisplay?: boolean;
+  manualCollectMode?: boolean; 
+  currentTestConfig?: string;
 }
 
 const SECONDS_TO_MILLISECONDS = 1000;
@@ -261,7 +266,12 @@ export class SuiteList extends React.Component<
         test={suite.test}
         defaultOpen={this.elements_count < 5 && !this.props.defaultClose}
         commonTestRunStatus={this.props.db_state.status}
+        moduleTechName={suite.name}
+        onTestsSelectionChange={this.props.onTestsSelectionChange}
+        selectedTests={this.props.selectedTests}
+        selectionSupported={this.props.selectionSupported}
         measurementDisplay={this.props.measurementDisplay}
+        manualCollectMode={this.props.manualCollectMode} 
       />
     );
   }